Moved the Card and Playerhand objects into the Deck class
