+++ DeckOfCards/Deck.java
@@ -5,7 +5,7 @@ import java.util.Random;
 /**
  * A standard deck of playing cards with various useful methods
  * @author Nathan Breunig
- * @version 1.5.1
+ * @version 1.01
  */
 public class Deck {
     private ArrayList<Card> myDeck = new ArrayList<Card>();
@@ -13,11 +13,9 @@ public class Deck {
     private ArrayList<PlayerHand> hands = new ArrayList<PlayerHand>();
     private Random rand = new Random();
     private int randNum;
-    private boolean acesHigh;
-    private boolean useJokers;
+    private boolean acesHigh, useJokers;
     private int jokerValue, heartValue, spadeValue, clubValue, diamondValue;
     private int aceValue, jackValue, queenValue, kingValue;
-    private boolean blackjackMode = false;
 
     /**
      * Basic constructor to set ace high, and to see if the deck should contain two joker cards
@@ -624,4 +622,278 @@ public class Deck {
     }
 
     public int getKingValue(){return kingValue;}
+
+    /**
+     * A card with a face value and suit
+     */
+    private class Card {
+        private String strCard;
+        private String suit;
+        private Deck cardDeck;
+
+        public Card(){}
+
+        /**
+         * Constructs a card with a face value and a suit
+         * @param card Card as string
+         * @param cardSuit Suit as String
+         */
+        public Card(Deck deck, String card, String cardSuit){
+            strCard = card;
+            suit = cardSuit;
+            cardDeck = deck;
+        }
+
+        /**
+         * Will get the face value of a card
+         * @return value of card
+         */
+        public int getValue() {
+            int value = 0;
+
+            if (strCard.equals("J")) {
+                value = cardDeck.getJackValue();
+            }else if (strCard.equals("Q")) {
+                value = cardDeck.getQueenValue();
+            }else if (strCard.equals("K")) {
+                value = cardDeck.getKingValue();
+            }else if (strCard.equals("Joker")){
+                value = cardDeck.getJokerValue();
+            }
+            else if (strCard.equals("A")) {
+                value = cardDeck.getAceValue();
+            }else{
+                value = Integer.parseInt(strCard);
+            }
+            return value;
+        }
+
+        public int getSuitValue(){
+            int value = 0;
+
+            if (suit.equals("Heart")){
+                value = cardDeck.getHeartValue();
+            }else if (suit.equals("Club")){
+                value = cardDeck.getClubValue();
+            }else if (suit.equals("Diamond")){
+                value = cardDeck.getDiamondValue();
+            }else if (suit.equals("Spade")){
+                value = cardDeck.getSpadeValue();
+            }
+            return value;
+        }
+
+        /**
+         * Will get the suit of a card
+         * @return Card suit as string
+         */
+        public String getSuit(){
+            return suit;
+        }
+
+
+        /**
+         * Gets the string of a card object
+         * @return The card as a string
+         */
+        public String toString(){
+
+            if (cardDeck.isAceHigh()) {
+                if (strCard.equals("14")) {
+                    return "A";
+                }
+            } else {
+                if (strCard.equals("1")){
+                    return "A";
+                }
+            }
+            if (strCard.equals("11")) {
+                return "J";
+            }else if (strCard.equals("12")) {
+                return "Q";
+            }else if (strCard.equals("13")) {
+                return "K";
+            } else{
+                return String.valueOf(strCard);
+            }
+        }
+    }
+
+    /**
+     * A players hand of cards
+     */
+    private class PlayerHand {
+        private ArrayList<Card> hand;
+
+        /**
+         * Constructs a PlayerHand
+         */
+        public PlayerHand(){
+            hand = new ArrayList<Card>();
+        }
+
+        /**
+         * Checking if a certain card (no matter suit) is in the hand
+         * @param card A card as string
+         * @return True if card is in hand, false otherwise
+         */
+        public boolean contains(String card){
+            for (int i = 0; i < hand.size(); i++){
+                if (hand.get(i).equals(card)){
+                    return true;
+                }
+            }
+            return false;
+        }
+
+        /**
+         * Checking if a certain card (suit included) is in the hand
+         * @param card A card as string
+         * @param suit A suit as string
+         * @return True if card is in hand, false otherwise
+         */
+        public boolean contains(String card, String suit){
+            for (int i = 0; i < hand.size(); i++){
+                if (hand.get(i).equals(card) && hand.get(i).getSuit().equals(suit)){
+                    return true;
+                }
+            }
+            return false;
+        }
+
+        /**
+         * Adds a card to the players hand
+         * @param card A Card object
+         */
+        public void add(Card card) {
+            hand.add(card);
+        }
+
+        /**
+         * Adds a card to the players hand
+         * @param pos The position to add it at (Ex. 1)
+         * @param card The card to add
+         */
+        public void add(int pos, Card card){
+            pos-=1;
+            hand.add(pos, card);
+        }
+
+        /**
+         * Adds a card to the front of the hand
+         * @param card The card to add
+         */
+        public void addFirst(Card card) {
+            hand.add(0, card);
+        }
+
+        /**
+         * Removes ALL cards of a certain value (no matter suit)
+         * @param card A card as a string
+         * @return A Card object
+         */
+        public Card remove(String card){
+            for (int i = 0; i < hand.size(); i++){
+                if (hand.get(i).toString().equals(card)){
+                    return hand.remove(i);
+                }
+            }
+            return null;
+        }
+
+        /**
+         * Removes a certain card (suit included)
+         * @param card A card as string
+         * @param suit A suit as string
+         * @return
+         */
+        public Card remove(String card, String suit){
+            for (int i = 0; i < hand.size(); i++){
+                if (hand.get(i).equals(card) && hand.get(i).getSuit().equals(suit)){
+                    return hand.remove(i);
+                }
+            }
+            return null;
+        }
+
+        /**
+         * Removes a random card from the hand
+         * @return A Card object
+         */
+        public Card removeRandom(){
+            Random rand = new Random();
+            if (hand.size() > 0){
+                return hand.remove(rand.nextInt(hand.size()));
+            }else{
+                throw new Error("No cards in hand");
+            }
+        }
+
+        /**
+         * Removes a card from the hand based on position
+         * @param pos Position of card in hand
+         * @return
+         */
+        public Card remove(int pos){
+            pos -= 1;
+            return hand.remove(pos);
+        }
+
+        /**
+         * Gets the size of the hand
+         * @return Size of hand as integer
+         */
+        public int getSize(){
+            return hand.size();
+        }
+
+        /**
+         * Gets a card from the hand based on position
+         * @param pos Position of card in hand
+         * @return
+         */
+        public Card get(int pos, boolean remove){
+            if (remove){
+                pos -=1;
+                return hand.remove(pos);
+            }else{
+                pos -=1;
+                return hand.get(pos);
+            }
+        }
+
+        public Card get(int pos){
+            pos-=1;
+            return hand.get(pos);
+        }
+
+        public Card getNext(boolean remove){
+            if (remove){
+                return hand.remove(0);
+            }else{
+                return hand.get(0);
+            }
+        }
+
+        /**
+         * Returns a simple string to visualize hand
+         * @return String of cards
+         */
+        public String toString(){
+            String string = "";
+
+            for (int i = 0; i < hand.size(); i++){
+                string += hand.get(i) + " ";
+            }
+
+            return string;
+        }
+
+        /**
+         * Clears player hand
+         */
+        public void clear(){
+            hand.clear();
+        }
+    }
 }
